% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antidiagnol.R
\name{antidiagnol}
\alias{antidiagnol}
\title{antidiagnol}
\usage{
antidiagnol(n)
}
\arguments{
\item{n}{integer}
}
\value{
Antidiagnol matrix of order n.
}
\description{
antidiagnol performs the creation of Back diagnol matrix.
}
\details{
An anti-diagonal matrix is a square matrix where all the entries are
zero except those on the diagonal going from the lower left corner to the upper right
corner entries are equal to 1.

In the first row, the last column will be 1 and all other entries are 0.

In second row, last but one column is 1 and others are 0 and so on.
}
\examples{
antidiagnol(4)
#0    0    0    1
#0    0    1    0
#0    1    0    0
#1    0    0    0
}
