% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_to_T.R
\name{base_to_T}
\alias{base_to_T}
\title{Base_to_T}
\source{
The Base sequences were obtained from \href{http://www.math.ntua.gr/~ckoukouv/}{Christos Koukouvinos}
}
\usage{
base_to_T(dat, x)
}
\arguments{
\item{dat}{is the frame containing basesequences to be exported}

\item{x}{integer (order of the base sequence)}
}
\value{
4 T-sequences of length of 2x+p.
}
\description{
internal function and it is not exported. It converts base sequences to T-Sequences.
}
\details{
dat - Internal dataset containing 4 sequences in long form with lentgh n+p,n+p,n,n. Using the 4 basequences,
the function creates 4 sequences of length 2n+p,2n+p,2n+p,2n+p.
T-Sequnces are usually used in creating matrices of Goethel Seidal array.
}
\references{
Goethals, J. M. and Seidel, J. J. (1967). Orthogonal matrices with zero diagnol. Canad. J. Math., 19, 259-264.
}
