% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseseq.R
\name{baseseq}
\alias{baseseq}
\title{baseseq}
\source{
The Base sequences were obtained from \href{http://www.math.ntua.gr/~ckoukouv/}{Christos Koukouvinos}
}
\usage{
baseseq(order)
}
\arguments{
\item{order}{integer}
}
\value{
Required Basesequences of order of x
}
\description{
Extracts the selection of Basesequences from internal dataset. Not exported.
}
\details{
Create Basesequence of given order from the internal dataset Basesequence
Base sequences are available in the internal table for order= 1:35
}
\references{
Goethals, J. M. and Seidel, J. J. (1967). Orthogonal matrices with zero diagnol. Canad. J. Math., 19, 259-264.
}
