% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdn_kronecker_matrix.R
\name{cdn_kronecker_matrix}
\alias{cdn_kronecker_matrix}
\title{cdn_kronecker_matrix}
\usage{
cdn_kronecker_matrix(r)
}
\arguments{
\item{r}{integer}
}
\value{
12 or NULL
}
\description{
Checks Hadamard Matrix can be constructed by multiplying 2 existing Hadamard matrix.
}
\details{
This function checks whether the Hadamard matrix  can be constructed as multiple of 2 Hadamard matrix.
Returns the value 12, if it is possible otherwise NULL is returned.
}
\examples{
cdn_kronecker_matrix(8)
#12
cdn_kronecker_matrix(12)
#NULL

}
