% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdn_miyamoto.R
\name{cdn_miyamoto}
\alias{cdn_miyamoto}
\title{cdn_miyamoto}
\usage{
cdn_miyamoto(order)
}
\arguments{
\item{order}{integer}
}
\value{
9 or NULL
}
\description{
Checks Hadamard Matrix can be constructed using Ehlich's method.
}
\details{
In Miyamoto construction, if q= n/4 and q is a prime or prime power
and q=1 (mod 4),then there exists an Hadamard Matrix of order n.
}
\examples{
cdn_miyamoto(20)
#q=5, is a prime number and q=1(mod 4).
#9
cdn_miyamoto(16)
#NULL
}
\references{
Miyamoto, M. (1991). A Construction of Hadamard matrices. J. Math. Phy., 12, 311-320.
}
\seealso{
\code{\link{had_miyamoto}} for Miyamoto's construction method.
}
