% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdn_PaleyIIprimepower.R
\name{cdn_PaleyIIprimepower}
\alias{cdn_PaleyIIprimepower}
\title{cdn_PaleyIIprimepower
checks Hadamard Matrix can be constructed using Paley II  method.}
\usage{
cdn_PaleyIIprimepower(order)
}
\arguments{
\item{order}{integer}
}
\value{
15 or NULL
}
\description{
cdn_PaleyIIprimepower
checks Hadamard Matrix can be constructed using Paley II  method.
}
\details{
In Paley II method,  q=order/2-1 and q is prime power and q=1 (mod 4) then
it retuns 15 otherwise NULL.
}
\examples{
cdn_PaleyIIprimepower(340)
#15
cdn_PaleyIIprimepower(64)
#NULL
}
\references{
Paley, R.E.A.C. (1933). On Orthogonal matrices. J. Combin. Theory, A 57(1), 86-108.
}
\seealso{
\code{\link{PaleyIIPrimePower}} for Paley construction method.
}
