% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdn_PaleyIprimepower.R
\name{cdn_PaleyIprimepower}
\alias{cdn_PaleyIprimepower}
\title{cdn_PaleyIprimepower
checks Hadamard Matrix can be constructed using Paley I method.}
\usage{
cdn_PaleyIprimepower(order)
}
\arguments{
\item{order}{integer}
}
\value{
14 or NULL
}
\description{
cdn_PaleyIprimepower
checks Hadamard Matrix can be constructed using Paley I method.
}
\details{
In Paley I method,  If q=order-1 and q is prime power and q=3 (mod 4) then
it retuns 14 otherwise NULL.
}
\examples{
cdn_PaleyI(28)
#14
cdn_PaleyI(16)
#NULL
}
\references{
Paley, R.E.A.C. (1933). On Orthogonal matrices. J. Combin. Theory, A 57(1), 86-108.
}
\seealso{
\code{\link{PaleyIPrimePower}} for Paley I construction method.
}
