% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_method.R
\name{Get_method}
\alias{Get_method}
\title{Get_method}
\usage{
Get_method(order)
}
\arguments{
\item{order}{integer}
}
\value{
Method name of the given order.
}
\description{
Get_method helps finding the given order of the matrix is constructed by which method.
}
\examples{
Get_method(92) # Williamson method
Get_method(24)
# Paley I
}
