% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GFADD.R
\name{GFADD}
\alias{GFADD}
\title{GFADD}
\usage{
GFADD(GFElem, p, r)
}
\arguments{
\item{GFElem}{integer (Can be obtained by calling GFELEM function)}

\item{p}{integer (a prime number)}

\item{r}{integer (a positive integer)}
}
\value{
A matrix of size p^r x p^r
}
\description{
Addition table of GF(P^r)
}
\details{
This function returns addition table of Galois field of order p^r. To use this function, Minimum function, elements of GF are required.
Minimum functions are available in internal dataset. Elements can be generated using GFELEM function.
}
\examples{

p<-3
r<-2
cardin<-p^2
mf<-subset(HadamardR:::minimumfunction,HadamardR:::minimumfunction$s==cardin)
MF<-mf$coeff
GFElem<-GFELEM(p,r,MF)
GFADD(GFElem,p,r)
#Addition Table of GF(9)

}
