% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GFM.R
\name{GFM}
\alias{GFM}
\title{GFM
GFM Generate Multiplication table of GF(p^r), where p is a prime power.}
\usage{
GFM(cardin)
}
\arguments{
\item{cardin}{integer}
}
\value{
Multiplication table of GF(p^r)
}
\description{
GFM
GFM Generate Multiplication table of GF(p^r), where p is a prime power.
}
\details{
This function returns Multiplication table of Galois field of order p^r. To use this function, Minimum function, elements of GF are required.
Minimum functions are available in internal dataset. Elements can be generated using GFELEM function.
}
\examples{
GFM(9)
##       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9]
##  [1,]    1    1    1    1    1    1    1    1    1
##  [2,]    1    3    4    5    6    7    8    9    2
##  [3,]    1    4    5    6    7    8    9    2    3
##  [4,]    1    5    6    7    8    9    2    3    4
##  [5,]    1    6    7    8    9    2    3    4    5
##  [6,]    1    7    8    9    2    3    4    5    6
##  [7,]    1    8    9    2    3    4    5    6    7
##  [8,]    1    9    2    3    4    5    6    7    8
##  [9,]    1    2    3    4    5    6    7    8    9
}
