% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GFPrimeAdd.R
\name{GFPrimeAdd}
\alias{GFPrimeAdd}
\title{GFPrimeAdd}
\usage{
GFPrimeAdd(p)
}
\arguments{
\item{p}{integer}
}
\value{
Addition Table of GF(p) in the form of matrix of order p x p.
}
\description{
GFPrimeAdd creates the addition Table for GF(p), where p is a prime number
}
\details{
If the elements of GF(p) are 0,1,..,p-1 then the (i,j)th element of matrix
returned is addition of (i-1)th and (j-1)th elements. The additions are
subjected to modulo p.
}
\examples{
GFPrimeAdd(5)
     #[,1] [,2] [,3] [,4] [,5]
#[1,]    0    1    2    3    4
#[2,]    1    2    3    4    0
#[3,]    2    3    4    0    1
#[4,]    3    4    0    1    2
#[5,]    4    0    1    2    3




}
