% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goethal-seidal_array.R
\name{goethals_seidel_array}
\alias{goethals_seidel_array}
\title{goethals_seidel_array}
\usage{
goethals_seidel_array(A = NA, B = NA, C = NA, D = NA)
}
\arguments{
\item{A}{matrix}

\item{B}{matrix}

\item{C}{matrix}

\item{D}{matrix}
}
\value{
goethals_seidel matrix of order 4n
}
\description{
goethals_seidel_array  performs the construction of Hadamard
matrix by Goethals-Seidel method
}
\details{
For this function requrires the four matrices, all the four matrix are Circulant matrices
same order.R is an antidiagonal matrix of the same order With which it should satisfy
the AA'+ BB'+ CC'+ DD'=4nI, where I is the identity matrix of order n.
This function returns matrix of order 4n where n is the order of the given matrices.
}
\references{
Goethals, J. M. and Seidel, J. J. (1967). Orthogonal matrices with zero
diagnol. Canad. J. Math., 19, 259-264.
}
