% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/had_cooper.R
\name{had_cooper}
\alias{had_cooper}
\title{had_cooper}
\source{
The williamson sequences and Turyn sequences were obtained from
\href{http://www.math.ntua.gr/~ckoukouv/}{Christos Koukouvinos}
}
\usage{
had_cooper(n)
}
\arguments{
\item{n}{integer (order of the matrix=876)}
}
\value{
Hadamard matrix of order n
}
\description{
had_cooper performs the construction of Hadamard matrix
by Cooper-Wallis method.In which combines the williamson matrices and T-sequences.
}
\references{
Cooper, J. and Wallis, W. D. (1972). A construction for Hadamard arrays. Bull. Austral. Math. 7, 269-278.
}
