% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/had_ehlich.R
\name{had_ehlich}
\alias{had_ehlich}
\title{had_ehlich}
\usage{
had_ehlich(x)
}
\arguments{
\item{x}{Integer (order of the Hadamard matrix)}
}
\value{
Hadamard matrix of order x
}
\description{
had_ehlich performs the construction of Hadamard matrix by Ehlich method
}
\details{
Ehlich (1965)'s construction method requires order of the Hadamard matrix must
be a of the form (n-1)^2. Conditions are (i)Order=(n-1)^2; (ii) n is a prime
or prime power and 3(mod 4); (iii) (n-2) must be a prime or prime power. In
case, if all the three conditions are satisfied,then function will return Hadamard
matrix of order x otherwise NULL.
}
\examples{
had_ehlich(36)
had_ehlich(20)
#NULL
}
\references{
Ehlich, H. (1965). Neue Hadamard-matrizen. Arch. Math., 16, 34-36.
}
