% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Had_method.R
\name{Had_method}
\alias{Had_method}
\title{Had_method}
\usage{
Had_method(order)
}
\arguments{
\item{order}{integer (order of the Hadamard matrix)}
}
\value{
Method number
}
\description{
Had_method  performs the give order of the matrix is constructed by which method.
}
\details{
If the method number retuns, if it

1      kronecker method (power of 2 only)

2      PaleyI

3      PaleyII

4      Ehlich method

5      Williamson method

6      Baumert-Hall method

7      Goethals-Seidel by using Base sequences

8      Goethals-Seidel by using Turyn sequences

9      Miyamoto method

10     Suplimentary Difference Sets

11     Cooper-Wallis method

12     Kronecker product method

13     Goethals-Seidel by using T sequences

14     Paley I Prime Power

15     Paley II Prime Power
}
\examples{
Had_method(92)   # "5"
Had_method(324)  # "4"
}
