% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/had_miyamoto.R
\name{had_miyamoto}
\alias{had_miyamoto}
\title{had_miyamoto}
\usage{
had_miyamoto(n)
}
\arguments{
\item{n}{integer (order of the matrix)}
}
\value{
Hadamard matrix of n
}
\description{
had_miyamoto function perform the construction of the Hadamard matrix by using the
Miyamoto method
}
\details{
If the q=n/4, and q be a prime power and q=1 (mod 4). If there is a exists of
Hadamard matrix of order q-1, then there exists an Hadamard matrix of order 4q.
If given order is not satisfied it returns NULL.
}
\examples{
had_miyamoto(20)
had_miyamoto(24) #NULL

}
\references{
Miyamoto, M. (1991). A Construction of Hadamard matrices. J. Math. Phy., 12, 311-320.
}
