% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/had_SDS.R
\name{had_SDS}
\alias{had_SDS}
\title{had_SDS}
\usage{
had_SDS(x)
}
\arguments{
\item{x}{integer (order of the matrix)}
}
\value{
Hadamard matrix of order x
}
\description{
had_SDS performs the construction of Hadamard matrix from SDS.
}
\details{
This function construct the Hadamard matrix of given order can be constructed using Suplementary Diffrence sets.
For given order the SDS set is not available it retuns NULL
If SDS is available it Returns Hadamard matrix of given order.

SDS are available for 103,127,151,163,181,191,239,251,463,631 in the internal table.
}
\examples{
\donttest{
had_SDS(412)
}
\donttest{
had_SDS(508)
}
}
\references{
Djokovic, D. Z. (1992a). Skew Hadamard matrices of order 4x37 and 4x39. J. Combin. Theory, A 61, 319-321.

Djokovic, D. Z. (1992b). Construction of some new Hadamard matrices. Bull. Austral. math. Soc., 45, 327-332.

Djokovic, D. Z. (1992c). Ten new Hadamard matrices of skew type. Publ.Electrotechnickog Fak., Ser. Mathematika, Univ. of Belgrade, 3, 47-59.

Djokovic, D. Z. (1992d). Ten Hadamard matrices of order 1852 of Goethals-Seidel type. Europ. J. Combinatorics, 13, 245-248.

Djokovic, D. Z. (1994a). Two Hadamard matrices of order 956 of Goethals-Seidel type. Combinatorica, 14(3), 375-377.

Djokovic, D. Z. (1994b). Five new Hadamard matrices of order skew type. Austral. J. Combinatorics, 10, 259-264.
}
