\name{HadamardR-package}
\alias{HadamardR-package}

\title{Hadamard Matrices}
\description{
A square matrix H of order n with entries +1 or -1 is called Hadamard Matrix, if HH'=nI, where I is Identity matrix. Necessary condition for Hadamard matrix of order n exists when n = 1,2 or 0(mod 4).
}
\details{
Using HadamardR, Hadamard Matrices of various orders can be generated. Out of 1250 possible Hadamard Matrices of order < 5000 (ignoring trivial orders 1 and 2), the construction methods are not known for 45 orders; 1158 orders are possible using Hadamard_Matrix() function. 47 Hadamard matrices not available in this package are as follows:
1336, 1432, 1940, 2212, 2264, 2292, 2316, 2488, 2740, 2776, 2864, 2872,
3140, 3352, 3476, 3544, 3620, 3684, 3704, 3708, 3820, 3832, 3880, 3892,
3896, 3928, 3972, 3980, 4044, 4120, 4152, 4184, 4268, 4296, 4304, 4344,
4396, 4404, 4432, 4528, 4572, 580, 4632, 4740, 4792, 4812, 4976
}


