% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Initial_row_SDS.R
\name{Initial_row_SDS}
\alias{Initial_row_SDS}
\title{Initial_row_SDS
Initial_row_SDS is an internal function.Not Exported.}
\usage{
Initial_row_SDS(i, j, v, n, r)
}
\arguments{
\item{i}{is the numeric vectors}

\item{j}{is the numeric vectors}

\item{v}{is the numeric vectors}

\item{n}{is the numeric vectors}

\item{r}{is the numeric vectors}
}
\value{
Intial rows of Matrix.
}
\description{
Initial_row_SDS
Initial_row_SDS is an internal function.Not Exported.
}
\details{
All inputs are numeric vectors of same length.
This function used in the COnstruction of Hadamard matrix by Supplementary Differences Sets
It converts the SDS sets into binary forms (+1 or -1).
}
\references{
Djokovic, D. Z. (1992a). Skew Hadamard matrices of order 4x37 and 4x39. J. Combin. Theory, A 61, 319-321.

Djokovic, D. Z. (1992b). Construction of some new Hadamard matrices. Bull. Austral. math. Soc., 45, 327-332.

Djokovic, D. Z. (1992c). Ten new Hadamard matrices of skew type. Publ.Electrotechnickog Fak., Ser. Mathematika, Univ. of Belgrade, 3, 47-59.

Djokovic, D. Z. (1992d). Ten Hadamard matrices of order 1852 of Goethals-Seidel type. Europ. J. Combinatorics, 13, 245-248.

Djokovic, D. Z. (1994a). Two Hadamard matrices of order 956 of Goethals-Seidel type. Combinatorica, 14(3), 375-377.

Djokovic, D. Z. (1994b). Five new Hadamard matrices of order skew type. Austral. J. Combinatorics, 10, 259-264.
}
