% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isprime.R
\name{is.prime}
\alias{is.prime}
\title{is.prime}
\usage{
is.prime(num)
}
\arguments{
\item{num}{integer}
}
\value{
TRUE or FALSE
}
\description{
is.prime check the given number is prime or not
}
\details{
if the given number is divisible any number other than 1 and itself it return NULL.
otherwise TRUE.
}
\examples{
is.prime(3)
#TRUE
is.prime(21)
#FALSE
}
