% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_primepower.R
\name{is.primepower}
\alias{is.primepower}
\title{is.primepower}
\usage{
is.primepower(p)
}
\arguments{
\item{p}{integer}
}
\value{
a and b where p=a^b and a is a prime number. Otherwise NULL
}
\description{
Checks whether given number is a prime power or not. Note that for a prime number, it would return NULL.
}
\details{
Returns a and b where p=a^b, otherwise NULL. Uses primeFactors() function of numbers package.
}
\examples{
is.primepower(2048)
#2 11
is.primepower(7)
#NULL
is.primepower(100)
#NULL
}
