% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_divisible.R
\name{is_divisible}
\alias{is_divisible}
\title{is_divisible}
\usage{
is_divisible(num, divisor)
}
\arguments{
\item{num}{integer}

\item{divisor}{integer}
}
\value{
num/divisor
}
\description{
is_divisible is internal function. Not exported.
}
\details{
it returns num/divisor value.
}
