% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jmat.R
\name{Jmat}
\alias{Jmat}
\title{Jmat}
\usage{
Jmat(n)
}
\arguments{
\item{n}{integer}
}
\value{
square matrix of order n
}
\description{
Jmat peforms the generation of unit matrix.
}
\details{
An J matrix is a square matrix where all the entries are one.
}
\examples{
Jmat(4)
#      [,1] [,2] [,3] [,4]
#[1,]    1    1    1    1
#[2,]    1    1    1    1
#[3,]    1    1    1    1
#[4,]    1    1    1    1
}
