% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kro_method.R
\name{kro_method}
\alias{kro_method}
\title{kro_method}
\usage{
kro_method(r)
}
\arguments{
\item{r}{integer (order of the matrix)}
}
\value{
r/2 or NULL.
}
\description{
kro_method internal function. Not exported.
}
\references{
Sylvester, J.J. (1967). Thoughts on orthogonal matrices, simultaneous sign-succession
and Tessellated pavements in two or more colours, with applications to Newton's rule,
ornamental Tie-work, and the theory of numbers. Phil. Mag.,34, 461-475.

Sylvester, J.J. (1968). Problem 2511. Math. Questions and solutions, 10, 74.

Hedayat, A. and Wallis, W.D. (1978). Hadamard Matrices and Their Application.Ann. Stat., 6, 1184-1238.
}
