% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kronecker_matrix.R
\name{kronecker_matrix}
\alias{kronecker_matrix}
\title{kronecker_matrix}
\usage{
kronecker_matrix(n)
}
\arguments{
\item{n}{integer (order of the matrix)}
}
\value{
Hadamard matrix of order "n"
}
\description{
kronecker_matrix
}
\details{
This function construct Hadamard matrix by multiple of 2 Hadamard matrix.
It Returns the Hadamard Matrix, if it is not possible NULL is returned.
}
\examples{
kronecker_matrix(8)
#      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8]
#[1,]    1    1    1    1    1    1    1    1
#[2,]    1   -1    1   -1    1   -1    1   -1
#[3,]    1    1   -1   -1    1    1   -1   -1
#[4,]    1   -1   -1    1    1   -1   -1    1
#[5,]    1    1    1    1   -1   -1   -1   -1
#[6,]    1   -1    1   -1   -1    1   -1    1
#[7,]    1    1   -1   -1   -1   -1    1    1
#[8,]    1   -1   -1    1   -1    1    1   -1
kronecker_matrix(12)
#NULL
}
\references{
Sylvester, J.J. (1967). Thoughts on orthogonal matrices, simultaneous sign-succession
and Tessellated pavements in two or more colours, with applications to Newton's rule,
ornamental Tie-work, and the theory of numbers. Phil. Mag.,34, 461-475.

Sylvester, J.J. (1968). Problem 2511. Math. Questions and solutions, 10, 74.

Hedayat, A. and Wallis, W.D. (1978). Hadamard Matrices and Their Application.Ann. Stat., 6, 1184-1238.
}
