% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method1_paleyII.R
\name{method1_paleyII}
\alias{method1_paleyII}
\title{method1_paleyII}
\usage{
method1_paleyII(n)
}
\arguments{
\item{n}{integer}
}
\value{
0 or (n/2)-1
}
\description{
method1_paleyII is internal function not exported.
}
\details{
this function checks q<- (n/2)-1, q is prime number and q = 1(mod 4). if it satisfy it returns q;
otherwise returns NULL.
}
\references{
Paley, R.E.A.C. (1933). On Orthogonal matrices. J. Combin. Theory, A 57(1), 86-108.
}
