% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method2_paleyII.R
\name{method2_paleyII}
\alias{method2_paleyII}
\title{method2_paleyII}
\usage{
method2_paleyII(n)
}
\arguments{
\item{n}{integer (order of the matrix)}
}
\value{
0 or (n/4)-1
}
\description{
method2_paleyII is internal function not exported.
}
\details{
this function checks q<- (n/4)-1, q is prime number and q = 1(mod 4). if it satisfy it returns q;
otherwise returns NULL.
}
\references{
Paley, R.E.A.C. (1933). On Orthogonal matrices. J. Combin. Theory, A 57(1), 86-108.
}
