% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miyamotoC.R
\name{miyamotoC}
\alias{miyamotoC}
\title{miyamotoC}
\usage{
miyamotoC(n)
}
\arguments{
\item{n}{integer (order of the matrix)}
}
\value{
q matrix
}
\description{
miyamotoC
}
\examples{
miyamotoC(20)
#      [,1] [,2] [,3] [,4] [,5]
#[1,]    0    1    1   -1   -1
#[2,]    1    0   -1   -1    1
#[3,]    1   -1    0    1   -1
#[4,]   -1   -1    1    0    1
#[5,]   -1    1   -1    1    0
}
\references{
Miyamoto, M. (1991). A Construction of Hadamard matrices. J. Math. Phy., 12, 311-320.
}
