% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nextElem.R
\name{nextElem}
\alias{nextElem}
\title{nextElem}
\usage{
nextElem(p1, MF, p, r)
}
\arguments{
\item{p1}{integer}

\item{MF}{integer}

\item{p}{integer}

\item{r}{integer}
}
\value{
A vector of order r, the coefficients of elements.
}
\description{
nextElem Generate next element of GF.
}
