% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normcol.R
\name{Normcol}
\alias{Normcol}
\title{Normcol
Normcol performs the Normalisation of column the given matrix.}
\usage{
Normcol(m)
}
\arguments{
\item{m}{Matrix}
}
\value{
Normalised matrix
}
\description{
Normcol
Normcol performs the Normalisation of column the given matrix.
}
\details{
For the given matrix of the first column of the all the -1 elements converting +1
without alter the property of the matrix.
}
\examples{
PaleyI(8)
#      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8]
#[1,]    1    1    1    1    1    1    1    1
#[2,]   -1    1   -1   -1    1   -1    1    1
#[3,]   -1    1    1   -1   -1    1   -1    1
#[4,]   -1    1    1    1   -1   -1    1   -1
#[5,]   -1   -1    1    1    1   -1   -1    1
#[6,]   -1    1   -1    1    1    1   -1   -1
#[7,]   -1   -1    1   -1    1    1    1   -1
#[8,]   -1   -1   -1    1   -1    1    1    1
Normcol(PaleyI(8))
#      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8]
#[1,]    1    1    1    1    1    1    1    1
#[2,]    1   -1    1    1   -1    1   -1   -1
#[3,]    1   -1   -1    1    1   -1    1   -1
#[4,]    1   -1   -1   -1    1    1   -1    1
#[5,]    1    1   -1   -1   -1    1    1   -1
#[6,]    1   -1    1   -1   -1   -1    1    1
#[7,]    1    1   -1    1   -1   -1   -1    1
#[8,]    1    1    1   -1    1   -1   -1   -1

}
