% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normrow.R
\name{Normrow}
\alias{Normrow}
\title{Normrow
Normcol performs the Normalisation of row the given matrix.}
\usage{
Normrow(m)
}
\arguments{
\item{m}{Matrix}
}
\value{
Normalised matrix
}
\description{
Normrow
Normcol performs the Normalisation of row the given matrix.
}
\details{
For the given matrix of the first row of the all the -1 elements converting +1
without alter the property of the matrix.
}
\examples{
PaleyII(12)
#      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12]
# [1,]    1    1    1    1    1    1    1   -1   -1    -1    -1    -1
# [2,]    1    1    1   -1   -1    1   -1    1   -1     1     1    -1
# [3,]    1    1    1    1   -1   -1   -1   -1    1    -1     1     1
# [4,]    1   -1    1    1    1   -1   -1    1   -1     1    -1     1
# [5,]    1   -1   -1    1    1    1   -1    1    1    -1     1    -1
# [6,]    1    1   -1   -1    1    1   -1   -1    1     1    -1     1
# [7,]    1   -1   -1   -1   -1   -1   -1   -1   -1    -1    -1    -1
# [8,]   -1    1   -1    1    1   -1   -1   -1   -1     1     1    -1
# [9,]   -1   -1    1   -1    1    1   -1   -1   -1    -1     1     1
#[10,]   -1    1   -1    1   -1    1   -1    1   -1    -1    -1     1
#[11,]   -1    1    1   -1    1   -1   -1    1    1    -1    -1    -1
#[12,]   -1   -1    1    1   -1    1   -1   -1    1     1    -1    -1
Normrow(PaleyII(12))
#       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12]
# [1,]    1    1    1    1    1    1    1    1    1     1     1     1
# [2,]    1    1    1   -1   -1    1   -1   -1    1    -1    -1     1
# [3,]    1    1    1    1   -1   -1   -1    1   -1     1    -1    -1
# [4,]    1   -1    1    1    1   -1   -1   -1    1    -1     1    -1
# [5,]    1   -1   -1    1    1    1   -1   -1   -1     1    -1     1
# [6,]    1    1   -1   -1    1    1   -1    1   -1    -1     1    -1
# [7,]    1   -1   -1   -1   -1   -1   -1    1    1     1     1     1
# [8,]   -1    1   -1    1    1   -1   -1    1    1    -1    -1     1
# [9,]   -1   -1    1   -1    1    1   -1    1    1     1    -1    -1
#[10,]   -1    1   -1    1   -1    1   -1   -1    1     1     1    -1
#[11,]   -1    1    1   -1    1   -1   -1   -1   -1     1     1     1
#[12,]   -1   -1    1    1   -1    1   -1    1   -1    -1     1     1
}
