% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PaleyIIPrimePower.R
\name{PaleyIIPrimePower}
\alias{PaleyIIPrimePower}
\title{PaleyIIPrimePower}
\usage{
PaleyIIPrimePower(order)
}
\arguments{
\item{order}{integer}
}
\value{
Hadamard matrix of the given order.
}
\description{
PaleyIIPrimePower
}
\details{
q=n/2-1, If there is an Hadamard matrix of order h>1, and q = 1 (mod 4) is a prime power,
then there exists an Hadamard matrix of order nh.
}
\examples{
PaleyIIPrimePower(20)
PaleyIIPrimePower(24)
}
\references{
Paley, R.E.A.C. (1933). On Orthogonal matrices. J. Combin. Theory, A 57(1), 86-108.
}
