% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PaleyIPrimePower.R
\name{PaleyIPrimePower}
\alias{PaleyIPrimePower}
\title{PaleyIPrimePower}
\usage{
PaleyIPrimePower(n)
}
\arguments{
\item{n}{integer}
}
\value{
Hadamard matrix
}
\description{
PaleyIPrimePower
}
\details{
let q = n-1 ,  and q = 3 (mod 4), q is the prime power, then obtained the Hadamard
matrix of order q+1.if input satisfies these condition it retuns Hadamard Matrix; otherwise
returns NULL.
}
\examples{
PaleyIPrimePower(28)
PaleyIPrimePower(28)
#NULL
}
\references{
Paley, R.E.A.C. (1933). On Orthogonal matrices. J. Combin. Theory, A 57(1), 86-108.
}
