% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ply1.R
\name{ply1}
\alias{ply1}
\title{ply1}
\usage{
ply1(q)
}
\arguments{
\item{q}{integer}
}
\value{
Hadamard matrix of order 2(q+1)
}
\description{
ply1 -internal function; not exported.
}
\references{
Paley, R.E.A.C. (1933). On Orthogonal matrices. J. Combin. Theory, A 57(1), 86-108.
}
