% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ply2.R
\name{ply2}
\alias{ply2}
\title{ply2}
\usage{
ply2(q)
}
\arguments{
\item{q}{integer}
}
\value{
Hadamard matrix of order 4(q+1)
}
\description{
ply2 is internal function and not exported
}
\references{
Paley, R.E.A.C. (1933). On Orthogonal matrices. J. Combin. Theory, A 57(1), 86-108.
}
