% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pow.R
\name{pow}
\alias{pow}
\title{pow}
\usage{
pow(n)
}
\arguments{
\item{n}{integer}
}
\value{
power of 2
}
\description{
pow functions finds the exponent of 2.
}
\details{
This function checks the given number is the power of 2 or not
If the given number is power of 2 it returns the exponent value; otherwise NULL is returned.
}
\examples{
pow(4)
# 2
pow(5)
#NULL
pow(6)
#NULL
}
