% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qhad2.R
\name{qhad2}
\alias{qhad2}
\title{qhad2}
\usage{
qhad2(p)
}
\arguments{
\item{p}{is the integer}
}
\value{
matrix of order p
}
\description{
qhad2 creats the Quadratic residues of the prime number.
}
\details{
The given input is prime number it retuns the matrix of order p.
if the input is not prime number it returns NULL.
}
\examples{
qhad2(7)
#      [,1] [,2] [,3] [,4] [,5] [,6] [,7]
#[1,]    0   -1   -1    1   -1    1    1
#[2,]    1    0   -1   -1    1   -1    1
#[3,]    1    1    0   -1   -1    1   -1
#[4,]   -1    1    1    0   -1   -1    1
#[5,]    1   -1    1    1    0   -1   -1
#[6,]   -1    1   -1    1    1    0   -1
#[7,]   -1   -1    1   -1    1    1    0
}
