% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QPrimepower.R
\name{QPrimePower}
\alias{QPrimePower}
\title{QPrimePower
QPrimePower creats the Quadratic residues of the prime number.}
\usage{
QPrimePower(cardin)
}
\arguments{
\item{cardin}{integer}
}
\value{
matrix of cardin x cardin
}
\description{
QPrimePower
QPrimePower creats the Quadratic residues of the prime number.
}
\details{
The given input is prime power it retuns the matrix of order cardin.
if the input is not prime number then it returns NULL.
}
\examples{
QPrimePower(9)
#      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9]
#[1,]    0    1   -1    1   -1    1   -1    1   -1
#[2,]    1    0    1   -1    1    1   -1   -1   -1
#[3,]   -1    1    0   -1    1   -1   -1    1    1
#[4,]    1   -1   -1    0    1   -1    1    1   -1
#[5,]   -1    1    1    1    0   -1    1   -1   -1
#[6,]    1    1   -1   -1   -1    0    1   -1    1
#[7,]   -1   -1   -1    1    1    1    0   -1    1
#[8,]    1   -1    1    1   -1   -1   -1    0    1
#[9,]   -1   -1    1   -1   -1    1    1    1    0
QPrimePower(36)
#NULL
}
