% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadprime.R
\name{quadprime}
\alias{quadprime}
\title{quadprime}
\usage{
quadprime(p)
}
\arguments{
\item{p}{integer}
}
\value{
squres
}
\description{
quadprime is a internal function not exported.
}
\details{
this function obtain Quadratic residues of GF. It returns squares of odd elements of GF
}
