% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T_seq.R
\name{T_seq}
\alias{T_seq}
\title{T_seq
T_seq performs the selection of the T sequences from dataset.internal function not exported.}
\source{
The Turyn sequences were obtained from \href{http://www.math.ntua.gr/~ckoukouv/}{Christos Koukouvinos}.
}
\usage{
T_seq(order)
}
\arguments{
\item{order}{integer}
}
\value{
Required Turyn sequences of order of x
}
\description{
T_seq
T_seq performs the selection of the T sequences from dataset.internal function not exported.
}
\details{
Create T sequences of given order from the internal dataset T_sequences

T-sequences are available for length of seq(1,73,2) and 83, 101 and 107 in the internal table.
}
\references{
Goethals, J. M. and Seidel, J. J. (1967). Orthogonal matrices with zero diagnol. Canad. J. Math., 19, 259-264.
}
\seealso{
\code{\link{had_goethals_T}} for Goethals-Seidel construction method using T-sequences.
}
