% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Turyn_seq.R
\name{Turyn_seq}
\alias{Turyn_seq}
\title{Turyn_seq
Turyn_seq performs the selection of the Turyn sequences from dataset.
It is internal function not exported.}
\usage{
Turyn_seq(order)
}
\arguments{
\item{order}{integer}
}
\value{
Required Turyn sequences of order of x
}
\description{
Turyn_seq
Turyn_seq performs the selection of the Turyn sequences from dataset.
It is internal function not exported.
}
\details{
Create Turyn sequences of given order from the internal dataset T_sequences

Turyn type-sequences are available for 28,30,34,36 in the internal table.
}
\references{
Goethals, J. M. and Seidel, J. J. (1967). Orthogonal matrices with zero diagnol. Canad. J. Math., 19, 259-264.
}
\seealso{
\code{\link{had_goethals_Turyn}} for Goethals-Seidel construction method using Turyn sequences.

#'
}
