% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T_to_base.R
\name{Turyn_to_T}
\alias{Turyn_to_T}
\title{Turyn_to_T
internal function. converts Turyn sequences to Base Sequences.}
\usage{
Turyn_to_T(dat1, order)
}
\arguments{
\item{dat1}{is the Turyn sequences subset exported from Tseq}

\item{order}{integer (order of the matrix)}
}
\value{
Basesequences of length of order
}
\description{
Turyn_to_T
internal function. converts Turyn sequences to Base Sequences.
}
\details{
dat - Internal dataset containing 4 sequences in long form with lentgh 2n-1, 2n-1, n, n. Using the 4 Turyn sequences,
the function creates 4 sequences of length n+p, n+p, n, n.
Base Sequnces are usually used in creating matrices of Goethel Seidal array.

Turyn type-sequences are available for 28,30,34,36 in the internal table.
}
\references{
Goethals, J. M. and Seidel, J. J. (1967). Orthogonal matrices with zero diagnol. Canad. J. Math., 19, 259-264.
}
\seealso{
\code{\link{had_goethals_Turyn}} for Goethals-Seidel construction method using Turyn sequences.
}
